plan weight become float + improved agent mind view with jFreeChart

+++ MetaCivTk3/src/civilisation/individu/decisionMaking/MaxWeightDecisionMaker.java
@@ -15,10 +15,10 @@ public class MaxWeightDecisionMaker extends DecisionMaker{
 		/* Select the new plan if there are no action to do */
 		if ((mind.getActionEnCours() == null))
 		{
-			int max = Integer.MIN_VALUE;
+			float max = Integer.MIN_VALUE;
 			NPlanPondere plan = null;
 			for (int i = 0 ; i < mind.getPlans().size(); i++) {
-				int weight = mind.getPlans().get(i).getPoids();
+				float weight = mind.getPlans().get(i).getPoids();
 				if (weight > max) {
 					max = weight;
 					plan = mind.getPlans().get(i);
